Move attributes tab registration to o11y root profile
