--- src/platform/plugins/shared/discover/public/context_awareness/profile_providers/observability/observability_base_document_profile/accessors/with_attributes_doc_viewer_tab.ts
@@ -1,52 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-import { i18n } from '@kbn/i18n';
-import type { DocumentProfileProvider } from '../../../../profiles';
-import { hasAnyFieldWithPrefixes } from '../../utils/has_any_field_with_prefixes';
-
-const attributesPrefixes = ['attributes.', 'scope.attributes.', 'resource.attributes.'];
-const hasAnyAttributesField = hasAnyFieldWithPrefixes(attributesPrefixes);
-
-const ATTRIBUTES_FEATURE_FLAG_ENABLED = false; // TODO replace with actual feature flag check
-export const withAttributesDocViewerTab = <
-  TProvider extends DocumentProfileProvider<TContext>,
-  TContext = TProvider extends DocumentProfileProvider<infer T> ? T : never
->(
-  provider: TProvider
-): TProvider => ({
-  ...provider,
-  profile: {
-    ...provider.profile,
-    getDocViewer: (prev, accessorParams) => (params) => {
-      const baseImpl = provider.profile.getDocViewer?.(prev, accessorParams);
-      const prevDocViewer = baseImpl?.(params) ?? prev(params);
-
-      return {
-        ...prevDocViewer,
-        docViewsRegistry: (registry) => {
-          if (hasAnyAttributesField(params.record) && ATTRIBUTES_FEATURE_FLAG_ENABLED) {
-            registry.add({
-              id: 'doc_view_obs_attributes_overview',
-              title: i18n.translate('discover.docViews.observability.attributesOverview.title', {
-                defaultMessage: 'Attributes',
-              }),
-              order: 9,
-              component: (props) => {
-                return 'Attributes Overview';
-              },
-            });
-          }
-
-          return prevDocViewer.docViewsRegistry(registry);
-        },
-      };
-    },
-  },
-});
