+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/observability/observability_root_profile/accessors/index.ts
@@ -9,3 +9,4 @@
 
 export { createGetAppMenu } from './get_app_menu';
 export { getDefaultAdHocDataViews } from './get_default_ad_hoc_data_views';
+export { getDocViewer } from './get_doc_viewer';
