+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/register_profile_providers.ts
@@ -28,7 +28,7 @@ import { createObservabilityRootProfileProvider } from './observability/observab
 import { createTracesDataSourceProfileProvider } from './observability/traces_data_source_profile';
 import { createDeprecationLogsDataSourceProfileProvider } from './common/deprecation_logs';
 import { createClassicNavRootProfileProvider } from './common/classic_nav_root_profile';
-import { createObservabilityDocumentProfileProviders } from './observability/observability_profile_provides';
+import { createObservabilityDocumentProfileProviders } from './observability/observability_profile_providers';
 
 /**
  * Register profile providers for root, data source, and document contexts to the profile profile services
